* Refactoring validation operations into its own controller and route * Adding summary operations to provide simple summarization of validation issues
